// packages/web/src/shared/api/generated/model/createPaymentTransactionDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Finnn API
 * Backend API for Finnn finance tracking
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePaymentTransactionDtoType } from "./createPaymentTransactionDtoType";
import type { NewPaymentCategoryDto } from "./newPaymentCategoryDto";

export interface CreatePaymentTransactionDto {
  accountId: string;
  amount: string;
  type: CreatePaymentTransactionDtoType;
  description?: string;
  date: string;
  categoryId?: string;
  newCategory?: NewPaymentCategoryDto;
}