// packages/web/src/shared/api/generated/model/transactionAccountDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Finnn API
 * Backend API for Finnn finance tracking
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionUserDto } from "./transactionUserDto";

export interface TransactionAccountDto {
  id: string;
  name: string;
  currency: string;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  ownerId?: string | null;
  /** @nullable */
  owner?: TransactionUserDto | null;
}